22:11

+++ src/Controller/Controller.java
@@ -57,15 +57,15 @@ public class Controller implements IController{
         M.saveInFile();
     }
 
-    public double get_temperatureInterieur() {
+    public float get_temperatureInterieur() {
         return M.get_temperatureInterieur();
     }
 
-    public double get_temperatureExterieur() {
+    public float get_temperatureExterieur() {
         return M.get_temperatureExterieur();
     }
 
-    public double get_temperaturePeltier() {
+    public float get_temperaturePeltier() {
         return M.get_temperaturePeltier();
     }
